Move stx enabled selectors to own file
