+++ app/selectors/preferencesController.ts
@@ -1,11 +1,6 @@
 import { createSelector } from 'reselect';
 import { PreferencesState } from '@metamask/preferences-controller';
 import { RootState } from '../reducers';
-import { isHardwareAccount } from '../../app/util/address';
-import { selectChainId, selectProviderConfig } from './networkController';
-import { NETWORKS_CHAIN_ID } from '../../app/constants/network';
-import Logger from '../util/Logger';
-import { swapsSmartTxFlagEnabled } from '../reducers/swaps';
 
 const selectPreferencesControllerState = (state: RootState) =>
   state.engine.backgroundState.PreferencesController;
@@ -115,52 +110,3 @@ export const selectSmartTransactionsOptInStatus = createSelector(
   (preferencesControllerState: PreferencesState) =>
     preferencesControllerState.smartTransactionsOptInStatus,
 );
-
-// TODO put this somewhere better
-export const ALLOWED_SMART_TRANSACTIONS_CHAIN_IDS = [
-  NETWORKS_CHAIN_ID.MAINNET,
-  NETWORKS_CHAIN_ID.GOERLI,
-  NETWORKS_CHAIN_ID.SEPOLIA,
-];
-
-export const getSmartTransactionsEnabled = (state: RootState) => {
-  const selectedAddress = selectSelectedAddress(state);
-  const addrIshardwareAccount = isHardwareAccount(selectedAddress);
-  const chainId = selectChainId(state);
-  const providerConfigRpcUrl = selectProviderConfig(state).rpcUrl;
-
-  const isAllowedNetwork =
-    ALLOWED_SMART_TRANSACTIONS_CHAIN_IDS.includes(chainId);
-
-  // E.g. if a user has a Mainnet Flashbots RPC, we do not want to bypass it
-  // Only want to bypass on default mainnet RPC
-  const canBypassRpc =
-    chainId === NETWORKS_CHAIN_ID.MAINNET
-      ? providerConfigRpcUrl === undefined
-      : true;
-
-  const smartTransactionsFeatureFlagEnabled = swapsSmartTxFlagEnabled(state);
-
-  const smartTransactionsLiveness =
-    state.engine.backgroundState.SmartTransactionsController
-      .smartTransactionsState?.liveness;
-
-  return Boolean(
-    isAllowedNetwork &&
-      canBypassRpc &&
-      !addrIshardwareAccount &&
-      smartTransactionsFeatureFlagEnabled &&
-      smartTransactionsLiveness,
-  );
-};
-
-export const getIsSmartTransaction = (state: RootState) => {
-  const isSmartTransactionsEnabled = getSmartTransactionsEnabled(state);
-  const smartTransactionsOptInStatus =
-    selectSmartTransactionsOptInStatus(state);
-
-  // return isSmartTransactionsEnabled && smartTransactionsOptInStatus;
-
-  // TODO remove this only for debugging
-  return true;
-};
